Added a simple maven configuration. Gradle does not compile anymore.
Made the member of VectorClassifier protected so that derivations of the class will be able to access the members.
Added a new constructor with full members, this way a custom tokenizer or stop word provider can be used.
Created a new stopword provider, which can read the default existing English words and allows new words to be added.